refactor: extract feedback cli module
